Extract io strategy
