+++ src/domain/core/src/entities/engine.rs
@@ -10,11 +10,8 @@
 use std::backtrace::Backtrace;
 use std::path::PathBuf;
 
-use ::serde::{Deserialize, Serialize};
-use ::serde_with::skip_serializing_none;
 use chrono::{DateTime, Utc};
 use internal_error::*;
-use opendatafabric::serde::yaml::*;
 use opendatafabric::*;
 use thiserror::Error;
 
@@ -39,31 +36,6 @@ pub trait IngestEngine: Send + Sync {
 ///////////////////////////////////////////////////////////////////////////////
 // Request / Response DTOs
 ///////////////////////////////////////////////////////////////////////////////
-
-#[skip_serializing_none]
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields, rename_all = "camelCase")]
-pub struct IngestRequestRaw {
-    #[serde(rename = "datasetID")]
-    pub dataset_id: DatasetID,
-    pub dataset_name: DatasetName,
-    pub input_data_path: PathBuf,
-    pub output_data_path: PathBuf,
-    #[serde(with = "datetime_rfc3339")]
-    pub system_time: DateTime<Utc>,
-    #[serde(default, with = "datetime_rfc3339_opt")]
-    pub event_time: Option<DateTime<Utc>>,
-    pub offset: i64,
-    #[serde(with = "SetPollingSourceDef")]
-    pub source: SetPollingSource,
-    #[serde(with = "DatasetVocabularyDef")]
-    pub dataset_vocab: DatasetVocabulary,
-    pub prev_checkpoint_path: Option<PathBuf>,
-    pub new_checkpoint_path: PathBuf,
-    #[serde(default, with = "datetime_rfc3339_opt")]
-    pub prev_watermark: Option<DateTime<Utc>>,
-    pub data_dir: PathBuf,
-}
 
 /// A request for ingesting new data into a root dataset.
 ///
