+++ src/infra/core/src/engine/mod.rs
@@ -9,9 +9,11 @@
 
 mod engine_config;
 mod engine_container;
+mod engine_io_strategy;
 mod engine_odf;
 mod engine_provisioner_local;
 mod engine_spark;
 
 pub use engine_config::*;
+pub use engine_io_strategy::*;
 pub use engine_provisioner_local::*;
